refactor: rename `auth` directory name to `login`
