+++ projects/common/.eslintrc
@@ -30,6 +30,7 @@
   "rules": {
     "import/no-extraneous-dependencies": 0,
     "import/extensions": 0,
-    "no-unused-vars": "warn"
+    "no-unused-vars": "warn",
+    "import/prefer-default-export": 0
   }
 }
